Rename test_admin.py for consistency.
