+++ src/xngin/apiserver/routers/auth/test_auth_dependencies.py
@@ -14,7 +14,7 @@ from xngin.apiserver import flags
 from xngin.apiserver.conftest import delete_seeded_users
 from xngin.apiserver.dwh.dwh_session import DwhSession
 from xngin.apiserver.routers.admin.admin_common import DEFAULT_NO_DWH_SOURCE_NAME
-from xngin.apiserver.routers.admin.test_admin import find_ds_with_name
+from xngin.apiserver.routers.admin.test_admin_api import find_ds_with_name
 from xngin.apiserver.routers.auth.auth_dependencies import (
     PRIVILEGED_TOKEN_FOR_TESTING,
     TESTING_TOKENS,
